Adapter Pattern
It contain example of Oil Business where the buyer uses the system which can measure the quantity of oil in gallon but the seller have system which measure the quantity of oil in liter. So a adapter is used in between which  can convert the quantity of oil in gallon to liter. And then the final cost is calculated in dollars.