Proxy Pattern
It contains an example of a confidential folder which can only be accessed by CEO and Manager. And if some employee having role as Developer try to access this file , a message will be shown that "You don't have permission to access this folder.". And to makethis whole system work there is a proxy folder which calls the real folder if the role of employee is CEO or Manager, if employee have some other role the message will be shown and hence employee will not be able to access the confidential folder.