Proxy Pattern
It contains an example of a secured folder which can only be accessed by CEO and Manager. And if some employee having role as Developer try to access this file , a message will be shown that "Sorry! You don't have permission to access Secured Folder.". And to make this whole system work there is a FolderProxy which decide which folder to call General Folder or Secured Folder based on the Role of employee.